feat: add ability to create arbitrary temp users
This commit refactors a few underlying edx-user set up functions to be
able to produce temporary users via `digitalId` instead of expecting a
sole user whose first name and last name match a specific query.

This feature is provided for both student and district users.  It also
includes a teardown task for deleting users.  Note: temporary users are
not properly activated and are not intended to be usable for log in.